refactor(split): mover función parse_filename a utils y eliminar implementación duplicada en split_dataset
